Start of chests and trying to fix stats
Created a simple chest inventory script that will randomly generate a subset of items from a given list. It currently only works for one chest at a time. Also started to figure out how the items would change stats and the many bugs that have come up due to it.