Fixed stats and loading of items
There is now code using a temp script for player stats that shows the stats changing correctly based on the items. Also had to change how the items were loaded as there were some issues at first with transferring the same item from a chest to the player's inventory. Most of the needed inventory components are now done or in a state where it is clear what needs to be changed.